Add files via upload
Weather Sensor Data Analysis with Random Forest Regression:
This case study is to test a structured data stream on the framework and apply a RandomForest regression to predict the air quality index in the city of Nottingham, which is defined by the Geographical coordinates (latitude = 52.950001 and longitude = -1.150000) of the location. The data stream source is Open Weather for Air Pollution API hourly forecast of polluting gases. Descriptions about the dataset parameters and API call can be found on the website. The reminder of the section demonstrated the implementation and outcome of the case study.